EDU-1279 Extract message-board component
