+++ src/components/pages/room.less
@@ -88,63 +88,6 @@
   font-style: italic;
 }
 
-.RoomPage-messageBoard {
-  margin-top: 60px;
-}
-
-.RoomPage-messageBoardPostingSection {
-  display: grid;
-  gap: 15px;
-  grid-auto-flow: row;
-}
-
-.RoomPage-messageBoardPostingSectionControls {
-  margin-left: auto;
-}
-
-.RoomPage-messageBoardMessages {
-  margin-top: 40px;
-  display: grid;
-  gap: 30px;
-  grid-auto-flow: row;
-}
-
-.RoomPage-messageBoardMessage {
-  border-radius: @edu-border-radius-base;
-  border: 1px solid @edu-border-color-base;
-}
-
-.RoomPage-messageBoardMessageHeadline {
-  display: flex;
-  justify-content: space-between;
-  padding: (@edu-content-padding / 2);
-  background-color: @edu-background-color-base;
-  border-top-left-radius: @edu-border-radius-base;
-  border-top-right-radius: @edu-border-radius-base;
-}
-
-.RoomPage-messageBoardMessageHeadlineIcons {
-  display: grid;
-  gap: 5px;
-  grid-auto-flow: column;
-  align-items: center;
-}
-
-.RoomPage-messageBoardMessageIcon {
-  display: flex;
-  justify-content: flex-end;
-  font-size: @edu-font-size-sm;
-  color: @edu-text-color-secondary;
-
-  &:hover {
-    cursor: help;
-  }
-}
-
-.RoomPage-messageBoardMessageText {
-  padding: @edu-content-padding;
-}
-
 .RoomPage-members {
   .m-card-wrapper;
 }
