+++ src/components/pages/room.yml
@@ -14,16 +14,6 @@ room:
   documentsEmptyStateSubtitle:
     en: Documents created within this room are exclusively available to room members and not accessible to the rest of the community.
     de: Die in diesem Raum erstellten Dokumente sind ausschließlich für die Raummitglieder abrufbar und nicht für den Rest der Community zugänglich.
-  messageBoardEmptyStateTitle:
-    en: There are no messages on the board
-    de: Auf dem Board sind keine Nachrichten
-  messageBoardEmptyStateSubtitle:
-    en: |
-      As room owner, you can post messages to the board in order to inform room members about various topics.
-      Members see the message board on the room page. They also receive a notification and, optionally, an email, when a new message is posted.
-    de: |
-      Als Raum-Eigentümer kannst du Nachrichten auf das Board stellen, um die Raum-Mitglieder über diverse Themen zu informieren.
-      Mitglieder sehen das Message-Board auf der Raum-Seite. Sie bekommen ebenfalls eine Benachrichtigung und - optional - eine E-Mail, sobald eine neue Nachricht gepostet wurde.
   membersEmptyStateTitle:
     en: You have not invited any members
     de: Du hast noch keine Mitglieder eingeladen
@@ -63,24 +53,6 @@ room:
   documentsSectionHeadline:
     en: Documents
     de: Dokumente
-  messageBoardSectionHeadline:
-    en: Message board
-    de: Message-Board
-  writeMessage:
-      en: Write message
-      de: Nachricht schreiben
-  postMessageButtonText:
-    en: Post message
-    de: Nachricht abschicken
-  postMessageCheckboxText:
-    en: Send email to members
-    de: E-Mail an Mitglieder schicken
-  postMessageCheckboxInfo:
-    en: An email will be immediately sent to all room members, notifying them that there is new information on the room message board.
-    de: Eine E-Mail wird unmittelbar an alle Raummitglieder geschickt, um diese darüber zu benachrichtigen, dass es neue Informationen auf dem Raum-Message-Board gibt.
-  messageEmailNotificationIconTooltip:
-    en: An email was sent to all room members when this message was posted.
-    de: Eine E-Mail wurde an alle Raum-Mitglieder geschickt, als diese Nachricht eingestellt wurde.
   overview:
     en: Overview
     de: Überblick
