+++ src/components/room/message-board.less
@@ -1,0 +1,60 @@
+.MessageBoard {
+  margin-top: 60px;
+}
+
+.MessageBoard-headline {
+  .m-section-headline;
+}
+
+.MessageBoard-messages {
+  margin-top: 40px;
+  display: grid;
+  gap: 30px;
+  grid-auto-flow: row;
+}
+
+.MessageBoard-message {
+  border-radius: @edu-border-radius-base;
+  border: 1px solid @edu-border-color-base;
+}
+
+.MessageBoard-messageHeadline {
+  display: flex;
+  justify-content: space-between;
+  padding: (@edu-content-padding / 2);
+  background-color: @edu-background-color-base;
+  border-top-left-radius: @edu-border-radius-base;
+  border-top-right-radius: @edu-border-radius-base;
+}
+
+.MessageBoard-messageHeadlineIcons {
+  display: grid;
+  gap: 5px;
+  grid-auto-flow: column;
+  align-items: center;
+}
+
+.MessageBoard-messageIcon {
+  display: flex;
+  justify-content: flex-end;
+  font-size: @edu-font-size-sm;
+  color: @edu-text-color-secondary;
+
+  &:hover {
+    cursor: help;
+  }
+}
+
+.MessageBoard-messageText {
+  padding: @edu-content-padding;
+}
+
+.MessageBoard-postingSection {
+  display: grid;
+  gap: 15px;
+  grid-auto-flow: row;
+}
+
+.MessageBoard-postingSectionControls {
+  margin-left: auto;
+}
