+++ src/components/room/message-board.yml
@@ -1,0 +1,29 @@
+messageBoard:
+  emptyStateTitle:
+    en: There are no messages on the board
+    de: Auf dem Board sind keine Nachrichten
+  emptyStateSubtitle:
+    en: |
+      As room owner, you can post messages to the board in order to inform room members about various topics.
+      Members see the message board on the room page. They also receive a notification and, optionally, an email, when a new message is posted.
+    de: |
+      Als Raum-Eigentümer kannst du Nachrichten auf das Board stellen, um die Raum-Mitglieder über diverse Themen zu informieren.
+      Mitglieder sehen das Message-Board auf der Raum-Seite. Sie bekommen ebenfalls eine Benachrichtigung und - optional - eine E-Mail, sobald eine neue Nachricht gepostet wurde.
+  headline:
+    en: Message board
+    de: Message-Board
+  writeMessage:
+      en: Write message
+      de: Nachricht schreiben
+  postMessageButtonText:
+    en: Post message
+    de: Nachricht abschicken
+  postMessageCheckboxText:
+    en: Send email to members
+    de: E-Mail an Mitglieder schicken
+  postMessageCheckboxInfo:
+    en: An email will be immediately sent to all room members, notifying them that there is new information on the room message board.
+    de: Eine E-Mail wird unmittelbar an alle Raummitglieder geschickt, um diese darüber zu benachrichtigen, dass es neue Informationen auf dem Raum-Message-Board gibt.
+  emailNotificationIconTooltip:
+    en: An email was sent to all room members when this message was posted.
+    de: Eine E-Mail wurde an alle Raum-Mitglieder geschickt, als diese Nachricht eingestellt wurde.
