+++ src/styles/main.less
@@ -62,6 +62,7 @@
 @import '../components/navigation-desktop.less';
 @import '../components/pages/user-profile.less';
 @import '../components/resource-info-cell.less';
+@import '../components/room/message-board.less';
 @import '../components/dashboard/rooms-tab.less';
 @import '../components/default-page-footer.less';
 @import '../components/default-page-header.less';
